+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 NOTES:
 - provider: Add a note about the debugging usage of the provider.
+- provider: Add a publish script for provider.
 
 ## 1.10.0 (September 30, 2020)